pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-pptx>=1.0.2",
    "sentence-transformers>=5.1.0",
    "unstructured>=0.18.14",
    "uvicorn>=0.35.0",
]
